gateway-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
supabase>=2.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic[email]>=2.7.0
ruff>=0.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
